+++ Mods ChoGGi/Fix Borked Transport Stutter/metadata.lua
@@ -1,11 +1,13 @@
 return PlaceObj("ModDef", {
-	"title", "Fix: Borked Transport Stutter v0.3",
-	"version", 3,
-  "saved", 1534680000,
+	"title", "Fix: Borked Transport Stutter v0.4",
+	"version", 4,
+  "saved", 1537358400,
 	"steam_id", "1440879916",
 	"id", "ChoGGi_BrokedTransportStutter",
 	"author", "ChoGGi",
-	"code", {"Script.lua"},
+	"code", {
+		"Code/Script.lua"
+	},
 	"image", "Preview.png",
 	"lua_revision", LuaRevision,
 	"description", [[RC Transports on a route have a certain tendency to get stuck and bog the game down (high speed feels like normal speed).