+++ Expanded Cheat Menu/metadata.lua
@@ -1,7 +1,7 @@
 return PlaceObj("ModDef", {
-	"title", "Expanded Cheat Menu v9.1 Test",
-	"version", 90,
-	"saved", 1537185600,
+	"title", "Expanded Cheat Menu v9.1",
+	"version", 91,
+	"saved", 1537617600,
 	"steam_id", "1411157810",
 	"id", "ChoGGi_CheatMenu",
 	"image", "Preview.png",